Adopt class based routes

+++ routes/web.php
@@ -1,5 +1,10 @@
 <?php
 
+use App\Http\Controllers\Admin;
+use App\Http\Controllers\BusinessController;
+use App\Http\Controllers\MapController;
+use Illuminate\Support\Facades\Route;
+
 /*
 |--------------------------------------------------------------------------
 | Web Routes
@@ -12,23 +17,22 @@
 */
 
 
-Route::get('/', 'MapController@index')->name('map');
-Route::get('/businesses/{id}', 'BusinessController@view')->name('map.business.view');
+Route::get('/', [MapController::class, 'index'])->name('map');
+Route::get('/businesses/{id}', [BusinessController::class, 'view'])->name('map.business.view');
 
 Route::prefix('admin')
     ->middleware('auth')
     ->middleware('can:accessAdmin')
-    ->namespace('Admin')
     ->group(function () {
-        Route::get('/', 'AdminController@index')->name('admin.index');
-        Route::get('business/validate-field', 'BusinessController@validateField')->name('admin.business.validate-field');
-        Route::get('business/from-submission/{id}', 'BusinessController@createFromSubmission')->name('admin.business.createFromSubmission');
-        Route::get('business/{id?}', 'BusinessController@edit')->name('admin.business.edit');
-        Route::post('business', 'BusinessController@create')->name('admin.business.create');
-        Route::put('business/{id}', 'BusinessController@update')->name('admin.business.update');
-        Route::delete('business/{id}', 'BusinessController@delete')->name('admin.business.delete');
-        Route::get('submissions', 'SubmissionController@index')->name('admin.submissions.index');
-        Route::get('submissions/{id}', 'SubmissionController@view')->name('admin.submissions.view');
-        Route::post('submissions/{id}', 'SubmissionController@update')->name('admin.submissions.update');
+        Route::get('/', [Admin\AdminController::class, 'index'])->name('admin.index');
+        Route::get('business/validate-field', [Admin\BusinessController::class, 'validateField'])->name('admin.business.validate-field');
+        Route::get('business/from-submission/{id}', [Admin\BusinessController::class, 'createFromSubmission'])->name('admin.business.createFromSubmission');
+        Route::get('business/{id?}', [Admin\BusinessController::class, 'edit'])->name('admin.business.edit');
+        Route::post('business', [Admin\BusinessController::class, 'create'])->name('admin.business.create');
+        Route::put('business/{id}', [Admin\BusinessController::class, 'update'])->name('admin.business.update');
+        Route::delete('business/{id}', [Admin\BusinessController::class, 'delete'])->name('admin.business.delete');
+        Route::get('submissions', [Admin\SubmissionController::class, 'index'])->name('admin.submissions.index');
+        Route::get('submissions/{id}', [Admin\SubmissionController::class, 'view'])->name('admin.submissions.view');
+        Route::post('submissions/{id}', [Admin\SubmissionController::class, 'update'])->name('admin.submissions.update');
     });
